pyproject.toml
[project]
name = "fine-tuning-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "huggingface-hub>=0.33.0",
    "hydra-core>=1.3.2",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "trl>=0.18.2",
    "wandb>=0.20.1",
]
